Move calculator to flow directory - additional-commodity-code
This calculator is specific to the additional-commodity-code flow and so should
be co-located.

* SmartAnswer::Calculators:CommodityCodeCalculator
  -> SmartAnswer::AdditionalCommodityCodeFlow::Calculator

* SmartAnswer::Calculators::CommodityCodeCalculatorTest
  -> SmartAnswer::AdditionalCommodityCodeFlow::CalculatorTest

Since these changes do not affect the behaviour and all the regression tests
pass, I've updated the checksums with the new location of the calculator.
