+++ test/data/additional-commodity-code-files.yml
@@ -1,5 +1,5 @@
 ---
-lib/smart_answer_flows/additional-commodity-code.rb: 0dbde7aa4f33eb96b5a336c9e9d5bdff
+lib/smart_answer_flows/additional-commodity-code.rb: 7e2fa71ebc57de67f78e935b968bbd7e
 test/data/additional-commodity-code-questions-and-responses.yml: 24243e9af7ed9c7aca5ed6adbfa0c9b7
 test/data/additional-commodity-code-responses-and-expected-results.yml: 3f80e1b92ec7731959b0f8033a4f43c8
 lib/smart_answer_flows/additional-commodity-code/additional_commodity_code.govspeak.erb: 4d11f8bf0bcd29764893c4a12349de17
@@ -15,5 +15,5 @@ lib/smart_answer_flows/additional-commodity-code/questions/how_much_sucrose_1.go
 lib/smart_answer_flows/additional-commodity-code/questions/how_much_sucrose_2.govspeak.erb: 69fa5acf376f4936de63f1bfb5775318
 lib/smart_answer_flows/additional-commodity-code/questions/how_much_sucrose_3.govspeak.erb: 6b3c81896c9981224fddf7996495ee40
 lib/smart_answer_flows/additional-commodity-code/questions/how_much_sucrose_4.govspeak.erb: 46ad361a17576dba3aa9041c7857d2bb
-lib/smart_answer/calculators/commodity_code_calculator.rb: 2ef1afca017d2688f4b30edf1442dcb4
+lib/smart_answer_flows/additional-commodity-code/calculator.rb: 017b6831d73db7c6607feeea2a2e8d6a
 lib/data/commodity_codes_data.yml: e18434a8a7b02f644e69ebe45f518772
